convert the omitted price_out_usd into FloatField.

+++ pharma/models.py
@@ -19,7 +19,7 @@ class UmutiEntree(models.Model):
     price_in = models.IntegerField(default=0)  #7: ayo Carton/plaquette yaranguwe
     price_out = models.IntegerField(default=0) #8: ayo plaquette tuyidandaza
     price_in_usd = models.FloatField(default=0)  #7: ayo Carton/plaquette yaranguwe muri usd
-    price_out_usd = models.IntegerField(default=0) #8: ayo plaquette tuyidandaza muri usd
+    price_out_usd = models.FloatField(default=0) #8: ayo plaquette tuyidandaza muri usd
     difference = models.IntegerField(default=0) #9: benefice
     quantite_initial = models.IntegerField(default=0) # izinjiye ubwambere
     quantite_restant = models.IntegerField(default=0) #10: plaquette zisigaye
@@ -43,7 +43,7 @@ class UmutiEntreeBackup(models.Model):
     price_in = models.IntegerField(default=0)  #7: ayo Carton/plaquette yaranguwe
     price_out = models.IntegerField(default=0) #8: ayo plaquette tuyidandaza
     price_in_usd = models.FloatField(default=0)  #7: ayo Carton/plaquette yaranguwe muri usd
-    price_out_usd = models.IntegerField(default=0) #8: ayo plaquette tuyidandaza muri usd
+    price_out_usd = models.FloatField(default=0) #8: ayo plaquette tuyidandaza muri usd
     difference = models.IntegerField(default=0) #9: benefice
     quantite_initial = models.IntegerField(default=0) # izinjiye ubwambere
     quantite_restant = models.IntegerField(default=0) #10: plaquette zisigaye
